Add request validation with fastify-openapi-glue by moving routes to a service wrapper
